requirements.txt
flask
flask-sqlalchemy
flask-bootstrap
flask-login
wtforms
flask-wtf
click